# openapi.lock.json
{
  "components": {
    "schemas": {
      "AdvancedStatsRead": {
        "properties": {
          "accuracy_pct": {
            "title": "Accuracy Pct",
            "type": "number"
          },
          "average_time_ms": {
            "title": "Average Time Ms",
            "type": "integer"
          },
          "average_time_ms_by_level": {
            "items": {
              "$ref": "#/components/schemas/LevelTimeStat"
            },
            "title": "Average Time Ms By Level",
            "type": "array"
          },
          "best_score": {
            "title": "Best Score",
            "type": "integer"
          },
          "best_streak_days": {
            "title": "Best Streak Days",
            "type": "integer"
          },
          "correct_attempts": {
            "title": "Correct Attempts",
            "type": "integer"
          },
          "current_streak_days": {
            "title": "Current Streak Days",
            "type": "integer"
          },
          "total_attempts": {
            "title": "Total Attempts",
            "type": "integer"
          },
          "total_score": {
            "title": "Total Score",
            "type": "integer"
          },
          "total_sessions": {
            "title": "Total Sessions",
            "type": "integer"
          }
        },
        "required": [
          "total_sessions",
          "total_attempts",
          "correct_attempts",
          "accuracy_pct",
          "total_score",
          "best_score",
          "average_time_ms",
          "current_streak_days",
          "best_streak_days",
          "average_time_ms_by_level"
        ],
        "title": "AdvancedStatsRead",
        "type": "object"
      },
...
      "DailyEvolutionPoint": {
        "properties": {
          "accuracy_pct": {
            "title": "Accuracy Pct",
            "type": "number"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "correct_attempts": {
            "title": "Correct Attempts",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "total_score": {
            "title": "Total Score",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "attempts",
          "correct_attempts",
          "accuracy_pct",
          "total_score"
        ],
        "title": "DailyEvolutionPoint",
        "type": "object"
      },
      "EvolutionRead": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/DailyEvolutionPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "days",
          "points"
        ],
        "title": "EvolutionRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LevelTimeStat": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "average_time_ms": {
            "title": "Average Time Ms",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "integer"
          }
        },
        "required": [
          "level",
          "attempts",
          "average_time_ms"
        ],
        "title": "LevelTimeStat",
        "type": "object"
      },
...
      "MyStatsRead": {
        "properties": {
          "accuracy_pct": {
            "title": "Accuracy Pct",
            "type": "number"
          },
          "average_time_ms": {
            "title": "Average Time Ms",
            "type": "integer"
          },
          "best_score": {
            "title": "Best Score",
            "type": "integer"
          },
          "correct_attempts": {
            "title": "Correct Attempts",
            "type": "integer"
          },
          "total_attempts": {
            "title": "Total Attempts",
            "type": "integer"
          },
          "total_score": {
            "title": "Total Score",
            "type": "integer"
          },
          "total_sessions": {
            "title": "Total Sessions",
            "type": "integer"
          }
        },
        "required": [
          "total_sessions",
          "total_attempts",
          "correct_attempts",
          "accuracy_pct",
          "total_score",
          "best_score",
          "average_time_ms"
        ],
        "title": "MyStatsRead",
        "type": "object"
      },
...
      "RankedQueueRequest": {
        "properties": {
          "level": {
            "anyOf": [
              {
                "maximum": 12.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          }
        },
        "title": "RankedQueueRequest",
        "type": "object"
      },
      "RankedQueueResponse": {
        "properties": {
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "opponent_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Username"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "RankedQueueResponse",
        "type": "object"
      },
      "RankedResolutionRead": {
        "properties": {
          "delta_points": {
            "default": 0,
            "title": "Delta Points",
            "type": "integer"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "points_after": {
            "title": "Points After",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "detail",
          "points_after"
        ],
        "title": "RankedResolutionRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Differentiation Bee",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/competitive/queue": {
      "post": {
        "operationId": "queue_ranked_competitive_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankedQueueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Queue Ranked",
        "tags": [
          "Competitive"
        ]
      }
    },
    "/competitive/queue/status": {
      "get": {
        "operationId": "queue_status_competitive_queue_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Queue Status",
        "tags": [
          "Competitive"
        ]
      }
    },
    "/competitive/resolve/{session_id}": {
      "post": {
        "operationId": "resolve_ranked_competitive_resolve__session_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedResolutionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Resolve Ranked",
        "tags": [
          "Competitive"
        ]
      }
    },
...
    "/stats/me": {
      "get": {
        "operationId": "read_my_stats_stats_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyStatsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read My Stats",
        "tags": [
          "Stats"
        ]
      }
    },
    "/stats/me/advanced": {
      "get": {
        "operationId": "read_my_advanced_stats_stats_me_advanced_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancedStatsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read My Advanced Stats",
        "tags": [
          "Stats"
        ]
      }
    },
    "/stats/me/evolution": {
      "get": {
        "operationId": "read_my_evolution_stats_me_evolution_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolutionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read My Evolution",
        "tags": [
          "Stats"
        ]
      }
    },
...
  }
}